Rename Macro.h to MiddlewareMacro to avoid duplication of name
